backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
aiofiles==23.2.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9 